requirements.txt
PyYAML
AWSIoTPythonSDK
argparse

